Add definitions of substring an subsequence
Subsequence is a generalisation of substring, suffix, and prefix. Finding the longest string which is a subsequence of two or more strings is known as the longest common subsequence problem.

Example: The string "anna" is a subsequence of the string "banana":


A substring of a string is a prefix of a suffix of the string, and equivalently a suffix of a prefix. If one string is a substring of another, it is also a subsequence, which is a more general concept.

Example: The string "ana" is a substring (and subsequence) of banana at two different offsets: